Introducing Py3D!!!
Py3D is just like p3dthon, but with a much cooler name.
Also most of the code is being rewriten it a much more pythonic
fassion. This is so other people can start to use it forrealz.

The first step is with the dump file reader and distro maker.

more to come,
cc